Development

+++ app/controllers/users_controller.rb
@@ -1,13 +1,10 @@
 class UsersController < ApplicationController
   skip_before_action :gate_keeper, only: [:create]
 
+  before_action :find_user, except: [:create]
+
   def show
-    @user = User.find(params[:id])
-    @user_nonos = ""
-    @user.nonos.each do |nono|
-      @user_nonos << nono.name + ", "
-    end
-    2.times {@user_nonos.chop!}
+    @user_nonos = @user.nonos.map {|nono| nono.name }.join(", ")
     @upcoming_parties = @user.upcoming_parties
     @past_parties = @user.past_parties.last(3)
   end
@@ -32,8 +29,7 @@ class UsersController < ApplicationController
   end
 
   def edit
-    @user = User.find(params[:id])
-    if current_user.id == @user.id
+    if is_current_user?(@user)
       @nonos = Nono.all
     else
       flash[:notice] = "You are not this user."
@@ -46,19 +42,28 @@ class UsersController < ApplicationController
       params[:user].delete(:password)
       params[:user].delete(:password_confirmation)
     end
-    user = User.find(params[:id])
-    user.update_attributes(user_params)
-    user.nono_users.destroy_all
-    nonos_hash = params[:user][:restrictions]
-    if nonos_hash
-      nonos_hash.each_key do |key|
-        user.nono_users.create(nono_id: key)
+    user = @user
+    user.assign_attributes(user_params)
+    if user.save
+      user.nono_users.destroy_all
+      nonos_hash = params[:user][:restrictions]
+      if nonos_hash
+        nonos_hash.each_key do |key|
+          user.nono_users.create(nono_id: key)
+        end
       end
+      redirect_to user_path(user)
+    else
+      flash[:notice] = user.errors.full_messages.join(". ")
+      redirect_to edit_user_path(user)
     end
-    redirect_to user_path(user)
   end
 
   private
+
+  def find_user
+    @user = User.find(params[:id])
+  end
 
   def user_params
     params.require(:user).permit(:email, :password, :password_confirmation, :name, :photo_url, :fave)
